+++ src/index.d.ts
@@ -54,6 +54,7 @@ interface JsonSchema {
     format?: "date-time" | "email" | "hostname" | "ipv4" | "ipv6" | "uri" | string;
     ref?: string;
     primary?: boolean;
+    final?: boolean;
 }
 
 interface SchemaJSON {