+++ src/rx-document.js
@@ -137,6 +137,18 @@ export class RxDocument {
         if (path.includes('.item.'))
             throw new Error(`cannot get observable of in-array fields because order cannot be guessed: ${path}`);
 
+        if (path === this.primaryPath)
+            throw RxError.newRxError('cannot observe primary path');
+
+        // final fields cannot be modified
+        if (this.collection.schema.finalFields.includes(path)) {
+            throw RxError.newRxError(
+                'final fields cannot be observed', {
+                    path
+                }
+            );
+        }
+
         const schemaObj = this.collection.schema.getSchemaByObjectPath(path);
         if (!schemaObj) throw new Error(`cannot observe a non-existed field (${path})`);
 
@@ -239,11 +251,24 @@ export class RxDocument {
     set(objPath, value) {
         if (typeof objPath !== 'string')
             throw new TypeError('RxDocument.set(): objPath must be a string');
+
+        // primary cannot be modified
         if (!this._isTemporary && objPath === this.primaryPath) {
             throw new Error(
                 `RxDocument.set(): primary-key (${this.primaryPath})
                 cannot be modified`);
         }
+
+        // final fields cannot be modified
+        if (!this._isTemporary && this.collection.schema.finalFields.includes(objPath)) {
+            throw RxError.newRxError(
+                'final fields cannot be modified', {
+                    path: objPath,
+                    value
+                }
+            );
+        }
+
         // check if equal
         if (Object.is(this.get(objPath), value)) return;